Create Sticker_script.ahk
Put your sticker image in the same folder has all of these scripts. Be sure to name the sticker image file sticker.png or change the name of the image in this script. Also be sure to change the URL for the sticker pack, you can find the URL for each pack by sharing the pack with another telegram user.